requirements.txt
langchain
google-generativeai
python-dotenv
requests
feedparser